Address code climate comments
