+++ app/src/androidTest/java/com/github/sdp/mediato/DatabaseTests/DataBaseTestUtil.java
@@ -1,6 +1,7 @@
 package com.github.sdp.mediato.DatabaseTests;
 
 import com.github.sdp.mediato.data.CollectionsDatabase;
+import com.github.sdp.mediato.data.LocationDatabase;
 import com.github.sdp.mediato.data.UserDatabase;
 
 public class DataBaseTestUtil {
@@ -12,6 +13,7 @@ public class DataBaseTestUtil {
   public static void useEmulator() {
     UserDatabase.database.useEmulator("10.0.2.2", 9000);
     CollectionsDatabase.database.useEmulator("10.0.2.2", 9000);
+    LocationDatabase.database.useEmulator("10.0.2.2", 9000);
   }
 
   /**
@@ -20,6 +22,7 @@ public class DataBaseTestUtil {
   public static void cleanDatabase() {
     UserDatabase.database.getReference().setValue(null);
     CollectionsDatabase.database.getReference().setValue(null);
+    LocationDatabase.database.getReference().setValue(null);
   }
 
 }
