+++ app/src/main/res/menu/bottom_menu.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-  e
   <item
       android:id="@+id/feed"
       android:title="Feed"
